requirements.txt
# FastAPI and Server
fastapi==0.103.1
uvicorn==0.23.2

# MongoDB
motor>=3.3.1
pymongo==4.5.0

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6

# Environment Variables
python-dotenv==1.0.0

# Data Validation
pydantic==2.4.2
email-validator==2.1.0

# CAD Processing
OCC-Core>=7.7.0
ezdxf>=1.0.0
numpy>=1.24.3

# Utilities
python-dateutil==2.8.2
pytz==2023.3
aiofiles==23.2.1

# Email
aiosmtplib==2.0.2
sendgrid==6.10.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Code Quality
black==23.11.0
flake8==6.1.0